+++ DerailValleyUniversalJato/Patches/LocomotionInputWrapper.cs
@@ -7,7 +7,7 @@ static class Patch_CrouchRequested
 {
     static void Postfix(ref bool __result)
     {
-        if (PlayerManager.Car != null && Main.settings.DisableCrouchWhenInTrainCar)
+        if (PlayerManager.Car != null && Main.settings != null && Main.settings.DisableCrouchWhenInTrainCar)
             __result = false;
     }
 }